requirements_deploy.txt
discord.py==2.3.2
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-bootstrap==3.3.7.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
